Remove test_snippets. Move snippets and generated code to snippets.py
